+++ src/ios/IntercomBridge.h
@@ -7,20 +7,19 @@
 - (void)registerIdentifiedUser:(CDVInvokedUrlCommand*)command;
 - (void)registerUnidentifiedUser:(CDVInvokedUrlCommand*)command;
 - (void)reset:(CDVInvokedUrlCommand*)command;
+
 - (void)setSecureMode:(CDVInvokedUrlCommand*)command;
 
 - (void)updateUser:(CDVInvokedUrlCommand*)command;
 - (void)logEvent:(CDVInvokedUrlCommand*)command;
+
+- (void)unreadConversationCount:(CDVInvokedUrlCommand*)command;
+
 - (void)displayMessageComposer:(CDVInvokedUrlCommand*)command;
-- (void)displayConversationsList:(CDVInvokedUrlCommand*)command;
-- (void)setVisibility:(CDVInvokedUrlCommand*)command;
-- (void)setPreviewPosition:(CDVInvokedUrlCommand*)command;
-- (void)setPreviewPadding:(CDVInvokedUrlCommand*)command;
+- (void)displayConversationsList:(CDVInvokedUrlCommand*)command ;
 
-- (void)setupAPN:(CDVInvokedUrlCommand*)command;
-
-- (void)setupGCM:(CDVInvokedUrlCommand*)command;
-- (void)openGCMMessage:(CDVInvokedUrlCommand*)command;
+- (void)setLauncherVisibility:(CDVInvokedUrlCommand*)command;
+- (void)setInAppMessageVisibility:(CDVInvokedUrlCommand*)command;
 
 - (void)registerForPush:(CDVInvokedUrlCommand*)command;